+++ src/main/net-placeholders/PlaceholderDifferenceEvaluator.cs
@@ -245,6 +245,15 @@ namespace Org.XmlUnit.Placeholder {
                 return EvaluateConsideringPlaceholders((string) controlDetails.Value,
                     (string) testDetails.Value, outcome);
 
+            }
+
+            // comparing QName content of attributes
+            else if (comparison.Type == ComparisonType.ATTR_VALUE && controlDetails.Value is XmlQualifiedName &&
+                     testDetails.Value is XmlQualifiedName)
+            {
+                return EvaluateConsideringPlaceholders((XmlQualifiedName) controlDetails.Value,
+                    (XmlQualifiedName) testDetails.Value, outcome);
+
             // "test document has no attribute but control document has"
             } else if (IsMissingAttributeDifference(comparison)) {
                 return EvaluateMissingAttributeConsideringPlaceholders(comparison, outcome);
@@ -336,6 +345,19 @@ namespace Org.XmlUnit.Placeholder {
                 return outcome;
             }
             return ComparisonResult.EQUAL;
+        }
+
+        private ComparisonResult EvaluateConsideringPlaceholders(XmlQualifiedName controlQName,
+            XmlQualifiedName testQName,
+            ComparisonResult outcome)
+        {
+            if (controlQName.Namespace == testQName.Namespace)
+            {
+                return EvaluateConsideringPlaceholders(controlQName.Name, testQName.Name,
+                    outcome);
+            }
+
+            return outcome;
         }
 
         private ComparisonResult EvaluateConsideringPlaceholders(string controlText, string testText,