+++ RELEASE_NOTES.md
@@ -2,9 +2,12 @@
 
 ## XMLUnit.NET 2.11.1 - /Not Released, yet/
 
+* placeholders can now also be used inside of the local part of `xsi:type` attributes.
+  PR [#49](https://github.com/xmlunit/xmlunit.net/pull/49)
+
 * PlaceholderDifferenceEvaluator would cause InvalidCastException for documents with
   differences in `xsi:type` attributes.
-  Equivalent of XMLUnit Java Issue [#276](https://github.com/xmlunit/xmlunit/issues/276)
+  Equivalent of XMLUnit Java Issue [#xmlunit/276](https://github.com/xmlunit/xmlunit/issues/276)
 
 * added readme files for the nuget packages.
   Issue [#46](https://github.com/xmlunit/xmlunit.net/issues/46).